Set expected type to dynamic for arguments to dynamic call
Summary:
When checking a call such as `$x->foo(arg)` on a receiver `$x` of type `dynamic`, set the "expected type" of the arguments to `dynamic`. This is important especially for lambda arguments, otherwise we have no information to use for unannotated parameters (see new test).

Because checking of parameters on dynamic calls is done with respect to "dynamic aware" subtyping, it's necessary to extend the implementation of expected types with a `coerce` parameter that is passed to the subtype function.

Differential Revision: D33533044

fbshipit-source-id: db0cc4c9ae8da027cb6eb7467a38dae4b5b0eea4