Introduce new kill IR and vasm ops
Summary:
This diff adds two IR ops: KillActRec and KillLoc, mirroring the existing KillIter. It also adds a new, sentinel vasm instruction, killeffects, which propagates IR memory effects to vasm.

An instruction with "kill" memory effects has implications for both performance and correctness. For performance, having more kills is generally good: they allow us to elide stores to the killed locations. But these ops may also be needed for correctness, if we're going to re-use those memory locations for a different purpose (e.g. converting a local to a stack variable). Today, the only cases where we re-interpret memory like that are when we enter or exit inlined frames. mofarrell and I would like to change that, allowing arbitrary locations (including frames) accessible from the stack pointer to be re-purposed as stack locations.

In order to re-purpose a location that we currently treat as an ActRec, local, or iterator, we use these IR ops to kill the value at that location from the IR's perspective. However, vasm-graph-color also uses memory effects to rematerialize spilled values, and it cannot pick up memory effects from an IR instruction that is lowered into an empty sequence of vasm instructions. As a result, we also need this new killeffects pseudo-instruction.

Reviewed By: mofarrell

Differential Revision: D33502143

fbshipit-source-id: 4fddb5854083ea93642c479fe49f65aff3833be9