union all parent members before doing override checks
Summary:
In large class hierarchies, there are a lot of diamond patterns which means that member override checks, as it's currently written, will check some members multiple times. To remediate this, this diff unions the sets of members of all parents before doing the override checks.

This is currently done only for properties and methods, not for constants and type constants which are still done per parent. We may include them in a later diff.

Reviewed By: andrewjkennedy

Differential Revision: D32956172

fbshipit-source-id: 089e9ed502dab166b5be247136b2c60e69a59474