Reject calls in l-value position
Summary:
Outlaw the use of function calls in l-value positions, such as
```
(foo($x))[3] = 'a';
```
or
```
(bar($y))[] = 45;
```
In the case of a copy-on-write array (such as `vec` or `dict`), it is definitely a bug, as the update or append "gets lost". In the case of a by-reference collection, it's still a bad code smell. Let's outlaw it in the type-checker.

Reviewed By: Wilfred

Differential Revision: D33029580

fbshipit-source-id: da1f0eff52f39638b1e975d859940e3c5f3dcd85