Kill rx/pure verify
Summary:
Coeffect local verification has been moved to the parser so that runtime and typechecker can share the implementation.

This makes the hhas pure verifier useless. As for RX verification, RX project is dead for the time being. If it is ever revived, the enforcement should be implemented in the parser not at the verifier since verifier errors are very unfriendly to the end users.

One argument to keeping this verifier step is for vanilla hhas files but this verifier is very incomplete and it is better to kill it and invest the time into the parser one for future changes.

Reviewed By: ricklavoie

Differential Revision: D33559727

fbshipit-source-id: 848d1b38bfeb3b319dea6e20c4b2c5c0a07011c8