fix fanout for static methods/props
Summary:
It seems like we've had this bug for a long time all the way back to beginning of this half https://fburl.com/code/xpxrnmdp. It's surprising that we didn't catch it until now. I suspect it's because previously we wouldn't have old decls and simply fan out to everything, which would hide this bug. Since we rolled out old decls on Tuesday, we now do have old decls, and so this bug manifests at a much higher rate.

This diff updates our shallow fanout to correctly construct Static Method and Static Prop deps when fanning out. As a refresher when we detect a method/prop M for some class C has changed, we construct the dep (Dep_Type, C::M) and then use our dep graph to figure out where to fan out to. However, if we use the wrong Dep_Type, like we've done with static methods, the fan out will be wrong.

Differential Revision: D33573187

fbshipit-source-id: 1fd75266159b7bcfd72112a5f03c0f4c669e51e5