fix label parsing w.r.t. autocompletion
Summary:
To allow autocompletion to work with the experimental syntax `$c->get#Name()`,
I enabled the parser to create a method call with the `()` part was missing,
to allow parsing `$c->get#NameAUTO332` and trigger autocompletion.

This should only be allow for autocompletion, ie. with the `AUTO332` suffix.
If not, it allows to get around the unstable feature protection if the method
has no other arguments.

This diff makes sure we only enable this special parsing with autocompletion is
running

Reviewed By: CatherineGasnier

Differential Revision: D33579296

fbshipit-source-id: e0d92cd59b108279530e85210058c54ba649606b