Enable logging of specific typing errors using `Typing_error.iter`
Summary:
Two errors, `invariant_violation` and `exception_occurred`, perform logging at the point the error is added. In order to replace these side effecting functions with use of `Primary.t` errors, this diff introduces a function, `Typing_error.iter`, which accepts a `Typing_error.t` along with two side effecting functions, `on_prim` and `on_snd`, which are called when `Primary.t` or `Secondary.t` errors are encountered when traversing the error.

Note also that the type of `Primary.to_error` and `Callback.apply` have been changed to return options in order to support the suppression of `Invariant_violation` error reporting. When errors are fully plumbed through typing / subtyping and part of the return value, we won't need to handle suppression in the evaluation of _any_ error and this can be removed.

Reviewed By: CatherineGasnier

Differential Revision: D33532466

fbshipit-source-id: 2a0f4d2a8dc3485888e7dce5db3b8ead97895632