Fix bug in fanout calculation for consts and typeconsts
Summary: Consts and typeconsts have their types copied into descendant classes in folded decl (rather than being stored in a separate heap as methods and properties are). If we use a method or property B::f which was inherited from A, we register a dependency both upon B::f and A::f, which is what allows us to avoid the more expensive use of recheck_descendants_and_their_member_dependents here. However, when using a const, because of the copy-down behavior in folded decl, we've never registered the dependency upon the origin class (A::f in this example). Since we don't register a dependency upon the const at its class of origin in this way, we must always take the more expensive path for consts here.

Reviewed By: bobrenjc93

Differential Revision: D33590165

fbshipit-source-id: 8056f89b41fac92779d06b20259fe1f5de768220