cgroup, rss in ProfileTypeCheck.get_stats
Summary:
I'd like to record cgroup memory stats since they're the ultimate arbiter, and rss-anon in case ocaml isn't releasing its memory properly. However as discussed earlier, although it's cheap to fetch these, it's slightly too costly to be done every single file.

This diff therefore fetches those things only at start of hh, start of typecheck, start of batch, and end of batch. (It'd be possible for me to invent a hh.conf flag for what percentage of files get this extra checking too, but that's a lot of engineering that I don't think is justified.)

Reviewed By: CatherineGasnier

Differential Revision: D33556151

fbshipit-source-id: f60120e683b4270d18c8e7e05c655fc02a83d4cf