Make IRBuilder restore the BCMarker when starting a block
Summary:
When starting a block IRBuilder retores the saved framestate into that block,
but does not update the marker to be consistent with that framestate.  Make it
update the marker to be consistent.

This was not generally an issue because when we start a bytecode in the new
block we will update the marker to be consistent with framestate.  However if
we start generating guards nothing will update the marker prior to those
instructions being genned.

Differential Revision: D33587095

fbshipit-source-id: d2036472729bb99154c3434f450b8d34ddfb8736