Add `Assert_in_current_decl` constructor to `Typing_error.t`
Summary:
There are several errors which are constructed by examining the `Pos_or_decl.t` position of the first 'reason' message of a `Secondary.t` error. If the position is in the current decl, determined using a global mutable variable, then the reason is converted to a claim. If not, a claim is added indicating there is a primary position bug.

This diff adds a new `Assert_in_current_decl` constructor to our `Typing_error.t` data type which accepts a `Secondary.t` error and `Pos_or_decl.ctx`. The `to_user_error` function is also updated to accept the current span when evaluating the error. This allows us to use the mutable variable when adding the error in `errors.ml`

Reviewed By: CatherineGasnier

Differential Revision: D33585093

fbshipit-source-id: ff508fdefc9fc6d4e44b1a516d61a7f947fd9097