make class element flags abstract
Summary: Instead of the flags of class_elt being a bare int, it is now Typing_defs_flags.ClassElt.t which is abstract. This required to port some of the logic of typing_defs related to xhp attributes to typing_defs_flags, which in turn required to define xhp_attr in its own file.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D33622382

fbshipit-source-id: bd08b8a335d589c55bd178abe94970c6d3fc44f4