Render both parts of a NamedLocal immediate
Summary:
mofarrell's "NamedLocal" immediate type separates the concepts of where a local is stored and what local name from the original source code applies to the value stored there. When we print these immediates in our bytecode disassembly, we should show both components. I've chosen the format $x0;"x1" to represent a local stored in the slot $x0 with the name "x1", since the name is just an arbitrary string label.

There are also NamedLocal immediates where the name is unused and has been DCE-d away. We render these as: $x0;_

This change got a bit complicated, because we'd also like to be able to parse the output of our disassembler - see as.cpp - and this format is produced by both HHVM and by HackC. Instead of modifying HackC's format, we modify as.cpp to accept both the old format (translating it to a NamedLocal where id == name) and the new format (potentially creating a NamedLocal where id != name). We add unit tests to cover the old format, the new format, and an invalid instance of the new format. The invalid instance is one with a NamedLocal, where the name doesn't appear in declvars.

One minor subtlety that came up is that, in order to keep using the current code to parse a local name, the separator between NamedLocal id and name has to be one of " \t\n#;>". That's kind of silly...I picked ';' but am open to changing it if it's a problem. I think it should be fine. Coeffects immediates are also ;-separated.

Reviewed By: mofarrell

Differential Revision: D33552747

fbshipit-source-id: df018fb60c081cb6343adeaea4a5a322ab2191fd