Fix stack offsets in FCall bytecodes
Summary: In irgen, if we do stack manipulation and then generate an IR op that may throw, we need to record the new stack offset. There's one place in the FCall bytecodes where we missed that: a dynamic call on a string. I'm a bit confused as to why pop() and push() don't just do this by default. Seems like it would save a lot of trouble.

Reviewed By: mofarrell

Differential Revision: D33648331

fbshipit-source-id: b63b40d67cc04ae023625675cbdec10c201cbad7