Lowerer: remove unnecessary generics
Summary:
The lowerer used to be generic across syntax pointer types. It always acts on the same types now, so use those types directly.

Change `Env` and `S` types to use the concrete token types, then replace `S<'a, T, V>` with `S<'a>` and `Env<'a, TF>` with `Env<'a>`.

Reviewed By: hgoldstein

Differential Revision: D33661604

fbshipit-source-id: 5af2357fa92bb037a94a0d947cabf17079c439c3