Remove already constrained template on FixedStringMapBuilder
Summary:
The name `FixedStringMapBuilder` is a bit misleading.
The template parameter `V` is already constrained to be a subtype of `std::vector::size_t` due to how we are setting the FSMap (String -> V) by `m_map[name] = m_list.size();`. This is not a generic builder as `FixedStringMap` is allowed to map to more data types than Slots.

Let's update the name and make this constraint explicit.

Reviewed By: ricklavoie

Differential Revision: D33649895

fbshipit-source-id: c420082f0dfdefe860e7e6a9449032fdf9045fad