Add `Assert_in_current_decl` combinator to `Typing_error.Reasons_callback.t`
Summary:
This diff adds a new combinator to reasons callbacks which lifts an error code and list of 'reason' messages to a `User_error.t` by inspecting the position of the first message and ensuring it is in the current decl. The combinator is analogous to that introduced in D33585093 (https://github.com/facebook/hhvm/commit/3d3c081ba46351e28d1e5d45dd2a2f6e65461727).

The two errors using this pattern have also be translated to use the new combinator

Reviewed By: CatherineGasnier

Differential Revision: D33620568

fbshipit-source-id: ac6d9b445a72c339a6d5fce6f13b90d56d9d3cf8