Let supportdyn functions be called with like types
Summary: When calling a function of type `supportdyn<t>` with arguments of type `u1, ..., un`, first check if the call is well typed with the function type `t`. If not, we want to try the call with `dynamic` as the function type. However, if `t` is a function type, then calling with `dynamic` is more permissive than we'd like. So instead if any of the `ui` are like types of the form `~ti = dynamic | ti`, we remove the `~` and check `ti` against the original parameter. This is semantically justified as long as all of the `ti <D: dynamic` since then all of the argument types are subtypes of dynamic. We also adjust the return type to be `~return_type`.

Reviewed By: andrewjkennedy

Differential Revision: D33189708

fbshipit-source-id: 00685abf0d397feebdc2b57d357e1321f39ce10e