Remove unused nontoplevel declarations flag
Summary:
A search indicates that the Hack.Lang.Phpism.DisableNonToplevelDeclarations and the hhvm.lang.phpism.disable_nontoplevel_declarations flags seems unused in hack and they are also absent from config files.

Hack also has no logic that is controlled by this option.

This flag is by default set to false in many of Hack's configs, even though nontoplevel declarations are illegal in Hack and have been for a while.

This diff removes the flag to help cleanup flags.

Reviewed By: fredemmott, Wilfred

Differential Revision: D33591454

fbshipit-source-id: 50149a2c931f07ab25c4a33d2e53b28cbd8c184c