do not check overriding a synthesized member
Summary:
Members marked as "synthesized" are members coming from a `require extends` (not require implements for now).

For example:
```
class A {
  public function foo() void {}
}

trait T {
  require extends A;
}

class B extends A {
  use T;
}
```
In this case, the decl for T has member `foo` inherited from A marked as synthesized.
When checking "parent" T for B, we check that B::foo <: T::foo. This isn't necessary because we've already checked the `require extends A` statement for B.

This diff skip override checks when the parent's member is marked as synthesized.

The benefit in terms of perf is twofolds:
- there are less members to union when unioning the parents' members
- in some cases we avoid redundant checks. The rest of this summary explains when.

In the example above, we wouldn't actually do the subtype check B::foo <: T::foo because we'd detect that foo's origin in T and B is the same: A.
Let's adapt the example so that foo's origin in B is B:
```
class A {
  public function foo() void {}
}

trait T {
  require extends A;
}

class B extends A {
  use T;

  public function foo(arraykey $_): void {}
}
```
This time we still wouldn't do the redundant check because we union the parents' members, and A::foo and T::foo are the same.

Let's adapt again so that T::foo and A::foo are different:
```
class C {
  public function foo(int $_): void {}
}

class A extends C {
  public function foo(arraykey $_): void {}
}

trait T {
  require extends C;
}

class B extends A {
  use T;
}
```
This time we'll check twice the following subtype relation: A::foo <: C::foo, once when checking A against C, the second time when checking B against T.

NB: There is a whole class of redundant checks:
- when the member being checked is inherited from an ancestor A, since that member has already been checked against all ancestors of A when checking A.
- when a class hierarchy isn't a tree but a DAG. `require extends` edges typically make hierarchies into non-tree DAGs.

Reviewed By: andrewjkennedy

Differential Revision: D33584389

fbshipit-source-id: a0537601c65e6dbab3883f687ba6714813238b86