refactor adding dependency when calling check_override
Summary:
Move the call to add_dependency inside check_override, since it's done each time we call it.

This also required to fix the `member_kind` that is passed when checking consistent constructors from "method" to "constructor", which in turn required to add a "consistent" flag to MemberKind.Constructor. I believe the logic is more correct this way because check_override does check a lot for member_kind.

Reviewed By: vsiles

Differential Revision: D33552668

fbshipit-source-id: c76e5d45ee472aac4eca00bf8b6fbf61e2174b79