no need to check dynamically callable for members that are not inherited
Summary:
The only case when a member is not inherited between a parent and a child is when the parent is an interface and the child is a concrete class. I believe that in this case there is no need to check for the member being dynamically callable in the parent because the the child would anyway need to provide an implementation for that member (next diff will actually raise an error in this case).

This diff also allows to make next diff's code cleaner.

Reviewed By: francesco-zappa-nardelli

Differential Revision: D33553014

fbshipit-source-id: 6f1196bb448a9fe8deb5fa7926f3b27dcfbb45cf