Homogenize ArrayData heap size tracking
Summary:
Many array implementations, like VanillaVec, MonotypeVec, MonotypeDict, and StructDict, track their heap size by storing their MemoryManager size index in their header. (To be precise, it's the high byte of m_aux16, and the header is packed together by a call to one of several packSizeIndexAndAuxBits helpers.)

Let's make this representation a hard constraint on arrays, in order to de-virtualize the "heapSize()" operation. We have to change VanillaDict, VanillaKeyset, EmptyMonotypeVec, EmptyMonotypeDict, and LoggingArray:
- For the former two, we split up the HashTable::reqAlloc method into HashTable::computeIndexFromScale and HashTable::reqAllocIndex, and change callers to use both methods to record the index in the header afterwards.
- For the latter three, the size index is a compile-time constant, but we still record the index in the header on allocation.

We check that we've covered all cases by adding new checkInvariants() methods to arrays that previously didn't use this approach for "heapSize()".

Reviewed By: mofarrell

Differential Revision: D33687367

fbshipit-source-id: ab21a554e7024bfc8100a1dec025886c5641aa08