remote payload creation
Summary:
This second diff implements the remote payload creation. In future diffs we will implement the actual spawning of the workers. For now we simply print out the command we intend to run on remote workers.

 ---

This will be a stack of diffs implementing a simpler (and hopefully faster) rewrite of hulk.

The primary idea is to rip out the complex & chatty state machine and opt for a simpler "pure" mechanism with larger, deterministic batches. In this process we'll also make some drastic simplifications

1) Replace tons of complex state machine logic including heartbeats, work stealing, etc. that is heavily coupled with the single node TC loop with effectively two function calls - spawn() + collect().
2) Drastically reduce the chattiness of the system. We will no longer waste time on blocking RPCs checking if small batches of work are done. We'll make num_workers RPC fetches in parallel once and only once.
3) Make replay testing locally easier since all remote worker invocations will now be "pure" and reproducible.

Differential Revision: D32747934

fbshipit-source-id: e38a7eba4609c79acbd509b66911da28416c8bf7