add hulk simple scaffolding to remote worker
Summary:
This fourth diff introduces scaffolding so that when the hulk_simple flag is on, the --check command to hh_server that simply downloads the payload from manifold, typechecks, and then uploads the results back to manifold.

 ---

This will be a stack of diffs implementing a simpler (and hopefully faster) rewrite of hulk.

The primary idea is to rip out the complex & chatty state machine and opt for a simpler "pure" mechanism with larger, deterministic batches. In this process we'll also make some drastic simplifications

1) Replace tons of complex state machine logic including heartbeats, work stealing, etc. that is heavily coupled with the single node TC loop with effectively two function calls - spawn() + collect().
2) Drastically reduce the chattiness of the system. We will no longer waste time on blocking RPCs checking if small batches of work are done. We'll make num_workers RPC fetches in parallel once and only once.
3) Make replay testing locally easier since all remote worker invocations will now be "pure" and reproducible.

Differential Revision: D32771860

fbshipit-source-id: d3154538c52893e7fb4cb02d4ac83ce4307b5bf2