disallow diamond import of traits that define generic properties
Summary:
As discussed in the "Diamond Inclusion of Traits" HIP (D32460483 - line 193), it may be unsound to importing a generic trait that defines a generic property at different types.  An example of the unsoundness is:

```
trait MyTrait<T> {
  public T $x;
}

trait T1 {
  use MyTrait<int>;

  public function getInt() : int { return $this->x; }
}

trait T2 {
  use MyTrait<string>;
}

<<__EnableMethodTraitDiamond>>
class C {
  use T1, T2;

  public function __construct() { $this->x = "hello"; }
}

<<__EntryPoint>>
function main() : void {
  echo (new C())->getInt();
}
```

Hack accepts this code, which is obviously unsound and correctly fatals in HHVM.

This diff fixes the unsoundness: Hack rejects the diamond import of a generic trait (as `MyTrait<T>`) whenever it defines generic properties (`$x` in the example).

*Remark*: in the HIP I argued that diamond import of all traits that define properties should be banned, independently from the actual type of the properties.  This is overly restrictive, and results in about 150 errors in WWW (most of which in perfectly legit code).

Reviewed By: vsiles

Differential Revision: D33533086

fbshipit-source-id: 3b39f2013a5c426a2425df76b159a2a40ac7f704