no call for borrow here
Summary:
this `x.borrow()` in
```
let rc = table.get(x.borrow()).and_then(...);
```
is superfluous.

we have that `T: Borrow<Q>` and `x:Q` that is, `x` is a borrowed form of `T`.

when we write `x.borrow()`, we are exercising the blanket `Borrow<Q>` that says a `Q` can be borrowed as a `Q`
```
impl<T: ?Sized> Borrow<T> for T {
    #[rustc_diagnostic_item = "noop_method_borrow"]
    fn borrow(&self) -> &T {
        self
    }
}
```
that is, `x.borrow()` is effectively a no-op

Differential Revision: D33742406

fbshipit-source-id: 53191b6ae0146509852c94ac76c6a2a3f40c015e