Rewrite Path to avoid memory allocations and copying
Summary:
After some digging the next performance bottleneck in taint requests
proved to be the representation of `Path`s. The existing implementation used a
`vector<Hop>` which was problematic for a few reasons.

First, these can get pretty large so we end up spending a lot of time in `memcpy`
when we move a value from the `Stack` to the `Heap` (or vice versa).

Secondly, due to the way the `Stack` and the `Heap` classes are written, even
operations that should normally be cheap (e.g. peeking at the top of the stack)
became quite expensive: we'd spend a lot of time in malloc/free due to the
temporary objects and copies being created.

Looking at those problems, and at our current usage of these data structures, I
realized we are mostly reading/copying them and not writing as much. And even
when we do write something, it's usually just to extend the `Path`. This
sounded like a good fit for a copy on write datastructure so I created a poor
man's copy on write structure using effectively a singly linked list.

Most operations in the interpreter hot path are now simple pointer copies or
null checks, and even when we do need to extend a `Path`, things are much
cheaper now. For fun, I also threw in the use of an `Arena` to cut down on
memory allocation overhead (and this ensures I don't have to worry about
freeing memory).

I would like feedback on whether this approach will be fine as we scale up to
more advanced taint analyses as e.g. I don't know if this assumption will hold
when we do joins and the like.

Reviewed By: dkgi

Differential Revision: D33760500

fbshipit-source-id: 326e36ad2d1cf3491de834075cfbf0e358b4d94a