Store `TaintedFunctionSet`s per request instead of globally
Summary:
Profiling showed a non-negligible amount of CPU being spent on locks
and accesses in the concurrent map we were using. We don't actually need that
level of concurrency and I think we can trade that performance off for some
memory (I hope).

We now create the `TaintedFunctionSet`s once and each request gets its own copy
(in the `State`). We can then use a different, more optimized implementation
for the evicting map as it doesn't have to worry about thread safety.

For kicks I implemented an optimization where at the end of the request we
write the cache back to the global configuration, which can be reused by other
requests. While this isn't perfect (each request may have different access
patterns) it's better than starting later requests off with an empty cache.

Note that due to the weird issues with how `Configuration` objects are loaded
at startup we need some extra null checks - I'm hoping I can nail down the
issue in future diffs. It doesn't seem to affect our performance too much right
now though as it's more of an initialization time thing.

If the memory tradeoff isn't worth it we can revert this later and spend more
CPU instead.

Reviewed By: dkgi

Differential Revision: D33760628

fbshipit-source-id: bfdef99fb2b31783025bb43bf3ee9808ac59dc00