Optimize regex handling
Summary:
When running requests locally, a significant amount of time was being
spent matching regexes. This diff optimizes things so regex matching no longer
show up on profiles.

I first moved to a faster regex implementation - re2. In particular re2
supports `Set`s of regexes so you can match against many regexes efficiently in
one go - which will become more important over time as we add more sources/sinks.

That was not enough, though, as we were still doing a lot of regex matching. I
optimized that by adding a 2 layer cache (see comments in code). I wrapped this
up in a helper class since the logic was the same for sources and sinks.
Bikeshedding welcome on the name

There is one awkward thing I'm leaving to future diffs: sometimes the
Configuration object can be accessed before it's fully initialized, so we need
some additional null checks. This shows in the `Configuration::{source,sink}`
methods and also in the initialization code (which I had to rewrite)

Reviewed By: dkgi

Differential Revision: D33760405

fbshipit-source-id: 492b2c460523060ea30d6d15e516d45639ba5d99