Take node by value in Conser::mk
Summary:
Currently, `mk` does the following:

- hash `x` via `table.get`
- if a value is present and can be upgraded to a strong reference, return it
- otherwise,
  - clone `x` via `x.to_owned()` (to be used as the hashtbl key)
  - clone `x` via `x.clone()` (to copy into the ConsedImpl which will be heap-allocated)
  - hash `x` again via `table.insert`

We could avoid the second hash by using the entry API, and we could avoid one of the clones by taking `x` by value. After this suggested change, `mk` does the following:

- hash `x` via `table.entry`
- if a value is present and can be upgraded to a strong reference, return it
- otherwise,
  - clone `x` via `x.to_owned()` in `make_rc` (referencing `o.key()` or `v.key()`, depending on whether `x` was present in the table)

I'm not sure if this is better overall. We don't have any benchmarks. But I gave it a try and I thought you might be interested in taking a look.

Differential Revision: D33674675

fbshipit-source-id: 5f1c4f645460558166c46fbc64e9f324cb7ec9e3