Remove use of AstProvider in TypingCheckJob
Summary: I think it's strange that the typechecker accepts a FileInfo and requests per-decl ASTs from an AST cache. It seems like it would be more straightforward to accept a full-file AST and iterate over it. If we do things that way, and use the DefsIterator introduced in the previous diff, we can entirely remove the fileinfo and nast modules in rupro.

Reviewed By: shayne-fletcher

Differential Revision: D33773063

fbshipit-source-id: 8de205c553c7516438f42a40da49b3ed8098bebd