Remove AstProvider's cache
Summary: We've been working on removing the AstProvider cache in hh_server, so I think it's best not to introduce one in rupro. The cache does not currently prevent redundant work: decls are parsed using the decl-parser (not the AST-parser), so each file is only parsed once immediately before typechecking.

Reviewed By: shayne-fletcher

Differential Revision: D33773120

fbshipit-source-id: 74f933a73f2a4a8fe7505b184d59fd125bd324cd