Change representation of Symbol to Box<str>
Summary: It is wasteful to store `String` here--any `Consed` value must be immutable (else violate `Conser` invariants), so the capacity usize inside `String` as well as the extra capacity in the backing buffer are wasted memory. Instead, store `Box<str>`, which is 16 bytes instead of 24 (being only a pointer and a length), and which won't contain excess space in the heap-allocated slice (if it was constructed via [`String::into_boxed_str`](https://doc.rust-lang.org/std/string/struct.String.html#method.into_boxed_str), which promises to drop excess capacity).

Reviewed By: edwinsmith

Differential Revision: D33672338

fbshipit-source-id: a6cf8801cd60038d44b4709bb796a3a4ed54704b