Make DB transactions implicit
Summary:
Previously, `AutoloadDB` exposed a `SQLiteTxn` object, representing an outstanding transaction, to callers. I'm planning on creating a MySQL implementation of `AutoloadDB`, which... won't be using `SQLiteTxn` objects. Based on the way the C++ MySQL client (Squangle) is written, based on the fact that, with a networked DB, transactions must be begun and committed asynchronously, and based on the fact that `AutoloadDB` doesn't actually *use* fine-grained transactions, I think it makes more sense to roll transactions into the `AutoloadDB` connection itself.

Now, whenever you get an AutoloadDB connection, you are always querying the same snapshot of the DB, for the lifetime of the connection.

Unfortunately, "the lifetime of the connection" is currently the lifetime of the thread, since `AutoloadDB` is implemented as a thread local. This caused `SymbolMapTest` to break, since most tests update the DB in a different thread than where we query it. I've fixed the test by adding `getDB().commit()` to `waitForDBUpdate()`, which refreshes the transaction.

Reviewed By: bdchik

Differential Revision: D33696872

fbshipit-source-id: d6af00c5d46eacf32cb1212835835dbd2da61903