Add ReleaseShallow and DecReleaseCheck IR-ops
Summary:
In addition to freeing memory, the Release method on array like types also inspects the elements so as to decrement the refcount for each of them. However, for arrays that contain non refcounted types/values, iterating through the elements is redundant.

This diff introduces a profile based optimization which replaces a DecRef which is likely to invoke the destructor of an array like with a pair of two new IR-ops DecReleaseCheck (which checks decrements the refcount and checks if the object should be released) and ReleaseShallow (frees memory associated with the object).

This diff only applies this optimization to a vector containing elements that aren't refcounted. Future changes will expand the scope of this optimization.

Differential Revision: D32731946

fbshipit-source-id: ea6c08e42a331ca2177d4899e69c2d19ad33ad28